chess-replay/requirements.txt
pandas
pycob
gunicorn
python-chess